EDITOR can now have arguments, and 'pin' can now be left empty
The pin field under subCAProvider was problematic because the API
doesn't return the pin, which means that the user have to give the
pin every time they want to patch the subCAProvider.

From now on, users will only need to set the pin if they intend to
configure the pin.